Refactor - moved ZuberasDiedDynamicValue to shared area
