+++ Mage.Sets/src/mage/sets/championsofkamigawa/DrippingTongueZubera.java
@@ -29,6 +29,8 @@
 package mage.sets.championsofkamigawa;
 
 import java.util.UUID;
+
+import mage.abilities.dynamicvalue.common.ZuberasDiedDynamicValue;
 import mage.constants.CardType;
 import mage.constants.Rarity;
 import mage.MageInt;
